fix syntax of Catalyst::Runtime version checks
A version requirement on a use statement must be a number or a vstring.
It cannot be a string. A string will be treated as an argument to
->import. Catalyst::Runtime doesn't have an import method, so the
argument would just be ignored. Future versions of perl are intending to
make passing arguments to an undefined import method an error.